feat: add server-side retention policy
Implement a server-side retention policy at the API endpoint, with a default
value of 7 days. When a user deletes an item (either a todo or a label),
the item will be soft deleted and a retention policy will be applied.
The item will then be completely removed from the database after 7 days.

This functionality can later be utilized to create a trashcan module,
allowing users to set custom retention policies.