feat: migrate LayoutBodyTagEffect to groupEffect
Transform the LayoutBodyTagEffect component into the useLayoutBodyTagClass
custom hook, integrating it into the GroupEffect components of LayoutApp
and LayoutHome.

Relocate the GroupEffect components from the 'effects/' directory to a
new 'groupEffects/' directory. Update module aliases in tsconfigs and
jest.configs accordingly.

Remove unnecessary components as part of the refactoring process.