feat: enhance layout header rendering and fix FOUC
Add a layer on top of the layout header for better alignment of the
header menu and background colors.

Address the Flash of Unstyled Content (FOUC) issue in the mobile
homeNavigation menu, which occurs when routing from the demo directory
to the Home page, by conditionally assigning the 'hidden' class.