feat: update on sections with additional SmoothTransition
Add the wrapper to the page index and add a section element to the section
components. Adjust the transition type on the 'Get started' button in homeHero
for better consistency. Also, add SmoothTransition components to the
homeStartToday section component to enhance user experience. The transition
rate for SmoothTransition is set to 0.7 for a slower transition trigger.