feat(CI): Add workflow for automatic develop branch creation
Introduce a new GitHub Actions workflow that triggers the creation
of a 'develop' branch upon closure of a pull request to the 'master' branch.

This workflow is based on the assumption that a 'develop' branch is deleted
automatically upon merging into 'master'. The integration of this workflow
ensures that a current 'develop' branch always exists within the repository.