feat: Separate the workflows of building and deploying docker
This commit separates the build-deploy-docker workflow into two separate
workflows: build-push and deploy. This improves the maintainability of
the project by making it easier to understand and update the workflows.
The deploy workflow will wait for the build-push workflow to be completed
before initiating.

Additionally, this commit removes the build-deploy-cloudrun workflow as it
is no longer in use.