feat(CI/CD): Refine Github Actions Workflows
Adjusts 'cleanup-gcr-resources' to retain only the three most recent
resources, a reduction from the previous setting of ten.

In 'docker-image-deploy', multiple flags are modified:

- Add `--min-instances=1` to enhance the cold start time for Google Cloud Run.
- Add `--no-cpu-throttling` to establish 'CPU always allocated', leading
to a 25% cost decrease in CPU and memory usage. Beneficial when a CPU is i
continuously running, such as when min instance is greater than 0.
- Disable `--cpu-boost` by commenting it out, considering it does not
offer extra benefits when the min-instances is above 0 and 'CPU always
allocated' is in place.