test: Extend LabelItemButtonContent Testing
Expand the scope of testing for the LabelItemButtonContent component by
introducing unit/integration tests. Utilizing a mockedLabelItem, located
under the '/__mock__/label/', allows for effective mocking of labelItem.

In order to facilitate accurate testing, a 'data-testid' attribute has been
introduced to the SvgIcon component. Recognizing that usage of 'data-testid'
is typically reserved for cases where no other testing options are viable,
it is necessary in this instance due to the SvgIcon being a decorative
component without other distinct attributes or elements available for testing.