feat: README Update
The README document has been modified to reflect recent changes.
The Container Registry badge has been removed and replaced with the Artifact
Registry badge in light of the deprecation of Container Registry as of
May 15, 2023.

An explanatory note has been included to clarify the rationale
behind this badge replacement. Further updates encompass minor modifications
to badge links and a rephrase of the introduction section for enhanced clarity.