test: Add unit test for LabelItemDropdown
The optionsLabelItemDropdown prop has been relocated into label.const for
better maintainability. A new unit test has been added for LabelItemDropdown.
This test does not run all the functional logics of each dropdown item, as
this is a unit test. The actual implementation and integration will be covered
within the integration test.