fix: Resolve Mobile Navigation Rendering Issue
Address an issue where mobileNavigation remained open upon route changes.
The cause was due to the `useInitialNavigation` hook having dependencies, which
led to state caching and consequently, the mobileNavigation persisting in the
open state.

As the purpose of `useInitialNavigation` is to set the initial state of
navigation for each render in the home layout, removing the dependencies
successfully resolves the issue.