feat: Add service-worker for static asset caching
Incorporate a service-worker to handle caching of static assets. The assets
being cached include files under the `/_next/static` directory and all
images. The aforementioned directory contains assets generated during the
build process.

To streamline cache versioning, the build process includes
the generation of a new version identifier which is embedded in
server-worker.js to invalidate the cache when necessary.