test: Add navigationButton component testing
Incorporate unit tests for the navigationButton component, evaluating both
the rendering process and the state updates upon button clicks. As part of
these tests, modifications have been made to navigationButton to embed
screen-reader only text within iconButton.

Additionally, tests for the searchBar component are deliberately omitted
at this time, as the component is currently serving as a placeholder and
does not possess its final implementation.