fix: Resolve client scrollbar inconsistency due to `w-screen`
The inclusion of `w-screen` led to the manifestation of a horizontal scrollbar
whenever a vertical scrollbar emerged, causing rendering inconsistency.
Resolution involved the removal of `w-screen` from the containers.

Padding was appended to the top and bottom sections of the underConstruction
component, allowing for the scrollbar to be displayed in advance.