chore(CI): Add workflow for repository vulnerability scanning
Add workflow to perform vulnerability scanning on the repository. This
includes scanning the dependencies using Trivy. The workflow is triggered
on pull requests and will cause the run to fail if the severity level is
HIGH or CRITICAL.

To enhance integration with other repositories, the workflow outputs the
vulnerability results as a table format instead of SARIF. SARIF format
can be used to integrate into Github Security, but Github Security only
available for Enterprise plans or Organization accounts.