feat: Add the button ui component to appRouter
Add the button component to the under `app/components/ui/` by moving the
button component into the appRouter. As the current process involves the
modification and transition of the pageRouter's component into
appRouter, there might be the duplication of the component across the
projects.

Additionally, the alias modular paths have been updated to fit the
current directories.