feat: Enhance `createConfigs` util function
`createConfigs` util now allows optional value of null or undefined for
defaultOptions properties. If a defaultOption's property is undefined or null,
it will not appear in the final object. Providing an empty object as
defaultOptions will result in no objects being passed to the final output,
while omitting defaultOptions entirely will result in a type error. This
functionality facilitates selective inclusion of properties in the options.